requirements.txt
# requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
Pillow
python-multipart